open account 1.13

+++ BankingClient/src/main/java/bankingclient/BankingApp.java
@@ -142,7 +142,7 @@ public class BankingApp {
             case "loan_account":
                 newAccount.setInterestRate(3);
                 balance = balance + ((balance / 100) * 2);
-                newAccount.setMonthlyRepayment(balance / 60);
+                newAccount.setMonthlyRepayment((int) balance / 60);
                 break;
             default:
                 break;
